chore(deps): bump server and sdk-go, upgrade to urfave v3

+++ action/hook/hook.go
@@ -10,7 +10,7 @@ type Config struct {
 	Action  string
 	Org     string
 	Repo    string
-	Number  int
+	Number  int64
 	Page    int
 	PerPage int
 	Output  string
